feat: add SPI and CS pin to InitilizationError
The SPI and chip select pin are returned in the InitilizationError on
the error paths in SDCard::new and related functions. The SPI and CS are
passed by value into those functions. The sucess path gives a means of
getting back ownership of SPI and CS through the SDCard::release method.
The InitilizationError::release method provides a similar mechanism for
the error path.