refactor: extract Response creational trait
Response provides a generic means of creating a possibly multi-byte
response. This is implemented for R1Response (with 0 extra bytes) and
for R3Response and R7Response (each with 4 extra bytes).

Extended the execute_command() function to a trait function that that is
implemented for every Response type. execute_command() deals with the
logic necessary for reading the R1 response, checking if it shows the
rest of the response has been truncated, reading the extra bytes (if
needed) and creating the final Response.